pyproject.toml
[project]
name = "sasha-golubkin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "blinker==1.7.0",
    "bs4>=0.0.2",
    "icecream>=2.1.4",
    "lxml>=5.3.1",
    "notebook>=7.3.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.29.0",
    "selenium-wire>=5.1.0",
    "undetected-chromedriver>=3.5.5",
    "webdriver-manager>=4.0.2",
]
